test-generator/requirements.txt
psycopg2-binary==2.9.6
Faker==18.11.2
python-dotenv==1.0.0